refactor: APIRequestProtocol 프로토콜 기본구현 생성
- execute, createBody 같은 메서드를 프로토콜 기본구현으로 만들어서, 프로토콜을 채택만해도 메서드를 사용할 수 있게 함
- 그리고 APIRequestProtocol 이 Parserable 프로토콜을 상속하게 하여 encode, decode 메서드 또한 사용할 수 있게 함
